Update coding_exercise.rb
When running this again found that the postcode was not being added to the string. Had done it this way originally but decided that it was redundant to put the postcode in a variable and add it to the string later on. 

Using   address += "#{address['postcode']}"  did function when I put it in here originally but seems unreliable for a reason I am uncertain of.